Move big to bigint. Since we're returning values with 18decimals they will always be ints. If we start returning/ storing values by val / 1e18 then we should move to BigDecimals

+++ v2/perps-v2/perps-subgraph/src/calculations.ts
@@ -1,7 +1,7 @@
-import { BigDecimal, BigInt } from '@graphprotocol/graph-ts';
+import { BigInt } from '@graphprotocol/graph-ts';
 
-export function calculateVolume(tradeSize: BigInt, lastPrice: BigInt): BigDecimal {
-  return tradeSize.times(lastPrice).div(BigInt.fromI32(10).pow(18)).abs().toBigDecimal();
+export function calculateVolume(tradeSize: BigInt, lastPrice: BigInt): BigInt {
+  return tradeSize.times(lastPrice).div(BigInt.fromI32(10).pow(18)).abs();
 }
 
 export function calculateLeverage(size: BigInt, lastPrice: BigInt, margin: BigInt): BigInt {
